Refactor to not not use LiveComponents when not neccessary
